Renamed the all movies reducer
