+++ resources/assets/js/reducers/movies.js
